:sparkles: 낭만우편함/낭만모음집 코드 리팩토링
