+++ src/main/java/aromanticcat/umcproject/service/nangmanLetterBoxService/NangmanCollectionService.java
@@ -1,14 +1,19 @@
 package aromanticcat.umcproject.service.nangmanLetterBoxService;
 
-import aromanticcat.umcproject.web.dto.nangmanLetterBox.NangmanLetterBoxResponseDTO;
+import aromanticcat.umcproject.web.dto.nangmanLetterBox.NangmanCollectionResponseDTO;
 
 import java.util.List;
 
 public interface NangmanCollectionService {
 
-    List<NangmanLetterBoxResponseDTO.PreviewBothResultDTO> findCollection(int page, int pageSize);
+    List<NangmanCollectionResponseDTO.PreviewBothResultDTO> findCollection(int page, int pageSize);
 
-    NangmanLetterBoxResponseDTO.BothResultDTO findCollectionDetails(Long nangmanLetterId);
+    NangmanCollectionResponseDTO.BothResultDTO findCollectionDetails(Long nangmanLetterId);
 
     void likeCollection(Long nangmanLetterId,String emojiType);
+
+    List<NangmanCollectionResponseDTO.PreviewBothResultDTO> getMyLetterList(Long userId, int page, int pageSize);
+
+    List<NangmanCollectionResponseDTO.PreviewBothResultDTO> getMyReplyList(Long userId, int page, int pageSize);
+
 }
