+++ src/main/java/aromanticcat/umcproject/service/nangmanLetterBoxService/NangmanLetterBoxServiceImpl.java
@@ -18,7 +18,6 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -100,61 +99,5 @@ public class NangmanLetterBoxServiceImpl implements NangmanLetterBoxService {
                 today.atStartOfDay(),
                 today.plusDays(1).atStartOfDay().minusSeconds(1)
         );
-    }
-
-    // 사용자 ID로 해당 사용자가 작성한 편지 목록 조회
-    @Override
-    @Transactional
-    public List<NangmanLetterBoxResponseDTO.PreviewLetterResultDTO> getMyLetterList(Long userId, int page, int pageSize){
-        Pageable pageable = PageRequest.of(page, pageSize);
-
-        Page<NangmanLetter> myLetterPage = nangmanLetterRepository.findByMemberId(userId, pageable);
-
-        List<NangmanLetter> myLetterList = myLetterPage.getContent();
-
-        return myLetterList.stream().map( myLetter -> NangmanLetterBoxConverter.toPreviewLetterResultDTO(myLetter))
-                .collect(Collectors.toList());
-
-    }
-
-
-    @Override
-    @Transactional
-    public NangmanLetterBoxResponseDTO.PreviewReplyResultDTO getPreviewReceivedReply(Long userId, Long nangmanLetterId){
-
-        //특정 편지 조회
-        NangmanLetter nangmanLetter = nangmanLetterRepository.findByMemberIdAndId(userId, nangmanLetterId);
-
-        if(nangmanLetter == null) {
-            throw new IllegalArgumentException("해당 사용자에게 권한이 없거나, 존재하지 않는 편지입니다.");
-        }
-
-        //특정 편지에 대한 답장이 있는지 확인
-        if (nangmanLetter.getHasResponse()) {
-
-            return NangmanLetterBoxConverter.toPreviewReplyResultDTO(nangmanLetter.getNangmanReply());
-        } else {
-            // 답장이 없는 경우 메세지를 담은 Optional.empty() 반환
-            throw new IllegalArgumentException("아직 답장이 없습니다.");
-        }
-
-    }
-
-    @Override
-    @Transactional
-    public  List<NangmanLetterBoxResponseDTO.PreviewBothResultDTO> getMyReplyList(Long userId, int page, int pageSize){
-
-        Pageable pageable = PageRequest.of(page, pageSize);
-
-        // 사용자가 답장한 목록 조회
-        Page<NangmanReply> myReplyPage = nangmanReplyRepository.findByMemberId(userId, pageable);
-
-        // 리스트로 변환
-        List<NangmanReply> myReplyList = myReplyPage.getContent();
-
-        // 각 답장(+ 연결된 편지)에 대해 미리보기로 생성
-        return myReplyList.stream()
-                .map(reply -> NangmanLetterBoxConverter.toPreviewBothResultDTO(reply.getNangmanLetter(), reply))
-                .collect(Collectors.toList());
     }
 }
